pyproject.toml
[tool.poetry]
name = "telegram-llm-bot"
version = "0.1.0"
description = ""
authors = ["ma2za <mazzapaolo2019@gmail.com>"]
readme = "README.md"
packages = [{ include = "telegram_llm_bot", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.5"
python-dotenv = "^1.0.0"
motor = "^3.3.1"
pyyaml = "^6.0.1"
langchain = "^0.0.292"
openai = "^0.28.0"
async-lru = "^2.0.4"
mmh3 = "^4.0.1"
pydantic-settings = "^2.0.3"
weaviate-client = "^3.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
